+++ src/com/reason/ide/console/BsToolWindowFactory.java
@@ -38,8 +38,8 @@ public class BsToolWindowFactory implements ToolWindowFactory, DumbAware {
         DefaultActionGroup group = new DefaultActionGroup();
         group.add(new ScrollToTheEndToolbarAction(console.getEditor()));
         group.add(new ClearLogAction(console));
-        group.add(new MakeAction());
-        group.add(new MakeWorldAction());
+        group.add(new BsMakeAction());
+        group.add(new BsMakeWorldAction());
 
         ActionToolbar toolbar = ActionManager.getInstance().createActionToolbar("left", group, false);
         toolbar.setTargetComponent(console.getComponent());
