+++ src/com/reason/ide/console/CompilerAction.java
@@ -1,7 +1,5 @@
 package com.reason.ide.console;
 
-import javax.swing.*;
-import org.jetbrains.annotations.NotNull;
 import com.intellij.execution.ui.ConsoleView;
 import com.intellij.execution.ui.ConsoleViewContentType;
 import com.intellij.openapi.components.ServiceManager;
@@ -14,14 +12,20 @@ import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiDocumentManager;
 import com.intellij.psi.PsiFile;
 import com.reason.Compiler;
+import com.reason.CompilerType;
 import com.reason.bs.Bucklescript;
 import com.reason.ide.ORCompilerManager;
+import org.jetbrains.annotations.NotNull;
+
+import javax.swing.*;
 
 abstract class CompilerAction extends DumbAwareAction {
 
     CompilerAction(@NotNull String text, @NotNull String description, @NotNull Icon icon) {
         super(text, description, icon);
     }
+
+    public abstract CompilerType getCompilerType();
 
     void doAction(@NotNull Project project, @NotNull CliType cliType) {
         Compiler compiler = ORCompilerManager.getInstance().getCompiler(project);
